Move routing into main app definition; make form a component
